Allowed string, bool, and enum proptypes to be interactive
