+++ src-docs/src/components/guide_section/guide_section.js
@@ -3,83 +3,16 @@ import PropTypes from 'prop-types';
 import { flatten } from 'lodash';
 
 import {
-  EuiCode,
   EuiCodeBlock,
   EuiErrorBoundary,
   EuiSpacer,
   EuiTab,
-  EuiTable,
-  EuiTableBody,
-  EuiTableHeader,
-  EuiTableHeaderCell,
-  EuiTableRow,
-  EuiTableRowCell,
   EuiTabs,
   EuiText,
-  EuiTextColor,
-  EuiTitle,
-  EuiLink
+  EuiTitle
 } from '../../../../src/components';
 
-function markup(text) {
-  const regex = /(#[a-zA-Z]+)|(`[^`]+`)/g;
-  return text.split(regex).map((token, index) => {
-    if (!token) {
-      return '';
-    }
-    if (token.startsWith('#')) {
-      const id = token.substring(1);
-      const onClick = () => {
-        document.getElementById(id).scrollIntoView();
-      };
-      return <EuiLink key={`markup-${index}`} onClick={onClick}>{id}</EuiLink>;
-    }
-    if (token.startsWith('`')) {
-      const code = token.substring(1, token.length - 1);
-      return <EuiCode key={`markup-${index}`}>{code}</EuiCode>;
-    }
-    return token;
-
-  });
-}
-
-const humanizeType = type => {
-  if (!type) {
-    return '';
-  }
-
-  let humanizedType;
-
-  switch (type.name) {
-    case 'enum':
-      if (Array.isArray(type.value)) {
-        humanizedType = type.value.map(({ value }) => value).join(', ');
-        break;
-      }
-      humanizedType = type.value;
-      break;
-
-    case 'union':
-      if (Array.isArray(type.value)) {
-        const unionValues = type.value.map(({ name }) => name);
-        unionValues[unionValues.length - 1] = `or ${unionValues[unionValues.length - 1]}`;
-
-        if (unionValues.length > 2) {
-          humanizedType = unionValues.join(', ');
-        } else {
-          humanizedType = unionValues.join(' ');
-        }
-        break;
-      }
-      humanizedType = type.value;
-      break;
-
-    default:
-      humanizedType = type.name;
-  }
-
-  return humanizedType;
-};
+import GuideProps from './guide_props';
 
 
 export class GuideSection extends Component {
@@ -140,131 +73,9 @@ export class GuideSection extends Component {
   }
 
   renderPropsForComponent = (componentName, component) => {
-    if (!component.__docgenInfo) {
-      return;
-    }
-
-    const docgenInfo = Array.isArray(component.__docgenInfo) ? component.__docgenInfo[0] : component.__docgenInfo;
-    const { _euiObjectType, description, props } = docgenInfo;
-
-    if (!props && !description) {
-      return;
-    }
-
-    const propNames = Object.keys(props);
-
-    const rows = propNames.map(propName => {
-      const {
-        description: propDescription,
-        required,
-        defaultValue,
-        type,
-      } = props[propName];
-
-      let humanizedName = (
-        <strong>{propName}</strong>
-      );
-
-      if (required) {
-        humanizedName = (
-          <span>
-            <strong>{humanizedName}</strong> <EuiTextColor color="danger">(required)</EuiTextColor>
-          </span>
-        );
-      }
-
-      const humanizedType = humanizeType(type);
-
-      const typeMarkup = markup(humanizedType);
-      const descriptionMarkup = markup(propDescription);
-      let defaultValueMarkup = '';
-      if (defaultValue) {
-        defaultValueMarkup = [ <EuiCode key={`defaultValue-${propName}`}>{defaultValue.value}</EuiCode> ];
-        if (defaultValue.comment) {
-          defaultValueMarkup.push(`(${defaultValue.comment})`);
-        }
-      }
-      const cells = [
-        (
-          <EuiTableRowCell key="name" header="Prop">
-            {humanizedName}
-          </EuiTableRowCell>
-        ), (
-          <EuiTableRowCell key="type" header="Type">
-            <EuiCode>{typeMarkup}</EuiCode>
-          </EuiTableRowCell>
-        ), (
-          <EuiTableRowCell key="defaultValue" header="Default" hideForMobile={!defaultValue}>
-            {defaultValueMarkup}
-          </EuiTableRowCell>
-        ), (
-          <EuiTableRowCell key="description" header="Note" isMobileFullWidth={true} hideForMobile={!propDescription}>
-            {descriptionMarkup}
-          </EuiTableRowCell>
-        )
-      ];
-
-      return (
-        <EuiTableRow key={propName}>
-          {cells}
-        </EuiTableRow>
-      );
-    });
-
-    const title = _euiObjectType === 'type' ?
-      <EuiCode id={componentName}>{componentName}</EuiCode> :
-      <EuiText>{componentName}</EuiText>;
-
-    let descriptionElement;
-
-    if (description) {
-      descriptionElement = (
-        <div key={`description-${componentName}`}>
-          <EuiText>
-            <p>{markup(description)}</p>
-          </EuiText>
-          <EuiSpacer size="m" key={`propsSpacer-${componentName}`} />
-        </div>
-      );
-    }
-
-    let table;
-
-    if (rows.length) {
-      table = (
-        <EuiTable className="guideSectionPropsTable" compressed key={`propsTable-${componentName}`}>
-          <EuiTableHeader>
-            <EuiTableHeaderCell>
-              Prop
-            </EuiTableHeaderCell>
-
-            <EuiTableHeaderCell>
-              Type
-            </EuiTableHeaderCell>
-
-            <EuiTableHeaderCell>
-              Default
-            </EuiTableHeaderCell>
-
-            <EuiTableHeaderCell>
-              Note
-            </EuiTableHeaderCell>
-          </EuiTableHeader>
-
-          <EuiTableBody>
-            {rows}
-          </EuiTableBody>
-        </EuiTable>
-      );
-    }
-
-    return [
-      <EuiSpacer size="m" key={`propsSpacer-${componentName}-1`} />,
-      <EuiTitle size="s" key={`propsName-${componentName}`}><h3>{title}</h3></EuiTitle>,
-      <EuiSpacer size="s" key={`propsSpacer-${componentName}-2`} />,
-      descriptionElement,
-      table,
-    ];
+    return (
+      <GuideProps key={componentName} componentName={componentName} component={component} />
+    );
   }
 
   renderProps() {
